Add check developerId functionality

+++ src/main/java/command/DeleteDeveloper.java
@@ -1,14 +1,16 @@
 package command;
 
+import services.CheckDevelopers;
 import services.DeveloperService;
 import view.View;
 
 import java.sql.SQLException;
 
-public class DeleteDeveloper implements Command{
+public class DeleteDeveloper implements Command {
     public static final String DELETE_DEVELOPER = "delete dev";
     private final View view;
     DeveloperService developerService = new DeveloperService();
+    CheckDevelopers checkDevelopers = new CheckDevelopers();
 
     public DeleteDeveloper(View view) {
         this.view = view;
@@ -22,18 +24,22 @@ public class DeleteDeveloper implements Command{
     @Override
     public void execute() {
         int id;
-        while (true) {
-            try {
-                view.write("Please, enter developer id to delete: ");
-                id = Integer.parseInt(view.read());
-                break;
-            } catch (NumberFormatException e) {
-                view.write("Invalid value. Use digits");
-            }
-        }
         try {
-            developerService.deleteDeveloper(id);
-            view.write("Developer with id " + id + " successfully deleted");
+            while (true) {
+                try {
+                    view.write("Please, enter developer id to delete: ");
+                    id = Integer.parseInt(view.read());
+                    if (checkDevelopers.IsDeveloperIdExists(id)) {
+                        developerService.deleteDeveloper(id);
+                        view.write("Developer with id " + id + " successfully deleted");
+                        break;
+                    } else {
+                        System.out.println("Developer id doesn't exists");
+                    }
+                } catch (NumberFormatException e) {
+                    view.write("Invalid value. Use digits");
+                }
+            }
         } catch (SQLException e) {
             e.getStackTrace();
         }
